Fix some test failures for Dart for the Web
`currentDartPackagePath` won't work for the web, where paths are
transformed when compiled to JavaScript.  Disable the test for
browsers.

The `sortWithKey` test did not show much difference when run in a
browser (possibly due to the overhead being more expensive), and
sometimes was very slightly slower than the normal `sort`.  Make
the test use `BigInt`s instead of `int`s to make comparisons more
expensive so that performance benefits from `sortWithKey` are more
pronounced.

+++ test/list_extensions_test.dart
@@ -21,9 +21,11 @@ void main() {
     test('Sorts efficiently', () {
       var trackedParseCallCount = 0;
 
-      int trackedParse(String string) {
+      BigInt trackedParse(String string) {
         trackedParseCallCount += 1;
-        return int.parse(string);
+
+        // [BigInt] is significantly more expensive than native [int].
+        return BigInt.parse(string);
       }
 
       var stopwatch = Stopwatch()..start();
